Move XVIZ format and type parsing functions to @xviz/io
- Preserves previously exported symbols, marked with deprecated comment
- Adds private method on XVIZData to obtain stats
