+++ modules/io/src/common/xviz-data.js
@@ -11,71 +11,18 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import {parseBinaryXVIZ, isBinaryXVIZ, getXVIZMessageType} from '@xviz/parser';
+/* global Buffer */
+/* eslint-disable complexity */
+import {
+  getDataContainer,
+  parseBinaryXVIZ,
+  isGLBXVIZ,
+  isJSONString,
+  getXVIZMessageType
+} from './loaders';
 import {XVIZMessage} from './xviz-message';
 import {TextDecoder} from './text-encoding';
 import {XVIZ_FORMAT} from './constants';
-
-/* global Buffer */
-
-// expected return value of null | binary | string | object
-function getDataContainer(data) {
-  if (data === null || data === undefined) {
-    return null;
-  }
-
-  if (data instanceof Buffer || data instanceof ArrayBuffer || ArrayBuffer.isView(data)) {
-    return 'binary';
-  }
-
-  // Cover string || object
-  return typeof data;
-}
-
-// Return true if the ArrayBuffer represents a JSON string.
-//
-// Search the first and last 5 entries for evidence of
-// being a JSON buffer
-function isJSONStringTypeArray(arr) {
-  let firstChar = arr.slice(0, 5).find(entry => entry >= 0x20);
-  let lastChars = arr.slice(-5);
-
-  // Buffer.slice() does not make a copy, but we need one since
-  // we call reverse()
-  if (lastChars instanceof Buffer) {
-    lastChars = Buffer.from(lastChars);
-  }
-
-  let lastChar = lastChars.reverse().find(entry => entry >= 0x20);
-
-  firstChar = String.fromCharCode(firstChar);
-  lastChar = String.fromCharCode(lastChar);
-
-  return (firstChar === '{' && lastChar === '}') || (firstChar === '[' && lastChar === ']');
-}
-
-// returns true if the input represents a JSON string.
-// Can be either string or Uint8Array
-//
-// Search the first and last 5 entries for evidence of
-// being a JSON buffer
-export function isJSONString(str) {
-  if (str instanceof Uint8Array) {
-    return isJSONStringTypeArray(str);
-  }
-
-  if (typeof str === 'object') {
-    return false;
-  }
-
-  const beginning = str.slice(0, 5).trim();
-  const end = str.slice(-5).trim();
-
-  return (
-    (beginning.startsWith('{') && end.endsWith('}')) ||
-    (beginning.startsWith('[') && end.endsWith(']'))
-  );
-}
 
 // Represents raw xviz data and
 // can create an XVIZMessage
@@ -87,7 +34,6 @@ export function isJSONString(str) {
 // - arraybuffer which is a JSON string
 // - JSON object
 // - arraybuffer which is a GLB
-
 export class XVIZData {
   constructor(data) {
     this._data = data;
@@ -116,8 +62,12 @@ export class XVIZData {
     return this._dataFormat;
   }
 
+  // In some cases this can be as expensive as a parse, so we do not
+  // load this unless asked for explicitly.
   get type() {
-    if (!this._xvizType) {
+    if (this._message) {
+      return this._message.type;
+    } else if (!this._xvizType) {
       const rawType = getXVIZMessageType(this._data);
       if (rawType) {
         const parts = rawType.split('/');
@@ -168,16 +118,19 @@ export class XVIZData {
         msg = JSON.parse(data);
         break;
       case XVIZ_FORMAT.OBJECT:
-        // TODO: what is the recursive case?
-        // see parse-stream-data-message.js
         msg = data;
         break;
       default:
         throw new Error(`Unsupported format ${this._dataFormat}`);
     }
 
-    this._message = new XVIZMessage(msg);
-    return this._message;
+    const xvizMsg = new XVIZMessage(msg);
+    if (xvizMsg.data) {
+      this._message = xvizMsg;
+      return this._message;
+    }
+
+    return null;
   }
 
   _determineFormat() {
@@ -188,16 +141,16 @@ export class XVIZData {
           data = data.buffer.slice(data.byteOffset, data.byteOffset + data.byteLength);
         }
 
-        if (isBinaryXVIZ(data)) {
+        if (isGLBXVIZ(data)) {
           this._dataFormat = XVIZ_FORMAT.BINARY_GLB;
-        }
+        } else {
+          if (data instanceof ArrayBuffer) {
+            data = new Uint8Array(data);
+          }
 
-        if (data instanceof ArrayBuffer) {
-          data = new Uint8Array(data);
-        }
-
-        if (isJSONString(data)) {
-          this._dataFormat = XVIZ_FORMAT.JSON_BUFFER;
+          if (isJSONString(data)) {
+            this._dataFormat = XVIZ_FORMAT.JSON_BUFFER;
+          }
         }
         break;
       case 'string':
