+++ modules/io/src/readers/xviz-base-reader.js
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import {isJSONString} from '../common/xviz-data';
+import {isJSONString} from '../common/loaders';
 
 // Base class that uses a Source to read file XVIZ file data
 export class XVIZBaseReader {
